pyproject.toml
[build-system]
requires = ["setuptools==75.8.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "developer-utility-toolkit"
version = "0.1.0"
description = "Security-focused modular developer multi-tool CLI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Developer Utility Toolkit Contributors" }
]
keywords = ["cli", "developer-tools", "security", "formatting", "conversion"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Utilities"
]
dependencies = [
  "typer==0.16.0",
  "defusedxml==0.7.1"
]

[project.optional-dependencies]
image = ["pillow==11.1.0"]
web = ["requests==2.32.4"]
xml = ["lxml==5.3.0"]
yaml = ["pyyaml==6.0.2"]
all = [
  "pillow==11.1.0",
  "requests==2.32.4",
  "lxml==5.3.0",
  "pyyaml==6.0.2"
]
dev = [
  "black==24.10.0",
  "ruff==0.9.6",
  "mypy==1.14.1",
  "pytest==8.3.4",
  "pytest-cov==6.0.0",
  "bandit==1.8.2",
  "pip-audit==2.7.3",
  "safety==3.2.14",
  "types-requests==2.32.0.20241016"
]

[project.urls]
Homepage = "https://github.com/your-org/developer-utility-toolkit"
Repository = "https://github.com/your-org/developer-utility-toolkit"
Issues = "https://github.com/your-org/developer-utility-toolkit/issues"

[project.scripts]
toolkit = "toolkit.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "N"]
ignore = ["S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
plugins = []

[tool.pytest.ini_options]
addopts = "--cov=toolkit --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
